[rtloader, pkg/collector] Delete handleCrashes method from rtloader Apart from error-handling, the method does not make use of the attributes of the rtloader class.
This commit:
- Removes the rtloader method
- Moves implementation into the handle_crashes function
- Makes errors be handled in the `make2`/`make3` style
