+++ pkg/collector/python/init_nix.go
@@ -9,6 +9,8 @@
 package python
 
 import (
+	"unsafe"
+
 	"github.com/DataDog/datadog-agent/pkg/util/log"
 
 	"github.com/DataDog/datadog-agent/pkg/config"
@@ -18,6 +20,7 @@ import (
 #cgo !windows LDFLAGS: -ldatadog-agent-rtloader -ldl
 
 #include <datadog_agent_rtloader.h>
+#include <rtloader_mem.h>
 */
 import "C"
 
@@ -31,8 +34,13 @@ func initializePlatform() error {
 			cCoreDump = 1
 		}
 
-		if C.handle_crashes(rtloader, C.int(cCoreDump)) == 0 {
-			log.Errorf("Unable to install crash handler, C-land stacktraces and dumps will be unavailable")
+		var handlerErr *C.char = nil
+		if C.handle_crashes(C.int(cCoreDump), &handlerErr) == 0 {
+			log.Errorf("Unable to install crash handler, C-land stacktraces and dumps will be unavailable: %s", C.GoString(handlerErr))
+			if handlerErr != nil {
+				// pyErr tracked when created in rtloader
+				C._free(unsafe.Pointer(handlerErr))
+			}
 		}
 	}
