+++ rtloader/include/datadog_agent_rtloader.h
@@ -306,7 +306,7 @@ DATADOG_AGENT_RTLOADER_API int has_error(const rtloader_t *);
 DATADOG_AGENT_RTLOADER_API const char *get_error(const rtloader_t *);
 #ifndef _WIN32
 
-/*! \fn int handle_crashes(const rtloader_t *, const int)
+/*! \fn int handle_crashes(const int, char** error)
     \brief Routine to install a crash handler in C-land to better debug crashes on RtLoader.
     \param rtloader_t A rtloader_t * pointer to the RtLoader instance.
     \param int A const integer boolean flag indicating whether dumps should be created
@@ -323,7 +323,7 @@ DATADOG_AGENT_RTLOADER_API const char *get_error(const rtloader_t *);
 
     Currently only SEGFAULT is handled.
 */
-DATADOG_AGENT_RTLOADER_API int handle_crashes(const rtloader_t *, const int);
+DATADOG_AGENT_RTLOADER_API int handle_crashes(const int, char** error);
 #endif
 
 // PYTHON HELPERS
