+++ rtloader/rtloader/api.cpp
@@ -9,6 +9,20 @@
 #    include <dlfcn.h>
 #endif
 
+#ifndef _WIN32
+// clang-format off
+// handler stuff
+#include <execinfo.h>
+#include <csignal>
+#include <cstring>
+#include <sys/types.h>
+#include <unistd.h>
+
+// logging to cerr
+#include <errno.h>
+// clang-format on
+#endif
+
 #include <iostream>
 #include <sstream>
 
@@ -323,14 +337,82 @@ void clear_error(rtloader_t *rtloader)
 }
 
 #ifndef WIN32
+core_trigger_t core_dump = NULL;
+
+static inline void core(int sig)
+{
+    signal(sig, SIG_DFL);
+    kill(getpid(), sig);
+}
+
+//! signalHandler
+/*!
+  \brief Crash handler for UNIX OSes
+  \param sig Integer representing the signal number that triggered the crash.
+  \param Unused siginfo_t parameter.
+  \param Unused void * pointer parameter.
+
+  This crash handler intercepts crashes triggered in C-land, printing the stacktrace
+  at the time of the crash to stderr - logging cannot be assumed to be working at this
+  poinrt and hence the use of stderr. If the core dump has been enabled, we will also
+  dump a core - of course the correct ulimits need to be set for the dump to be created.
+  The idea of handling the crashes here is to allow us to collect the stacktrace, with
+  all its C-context, before it unwinds as would be the case if we allowed the go runtime
+  to handle it.
+*/
+#    define STACKTRACE_SIZE 500
+void signalHandler(int sig, siginfo_t *, void *)
+{
+    void *buffer[STACKTRACE_SIZE];
+    char **symbols;
+
+    size_t nptrs = backtrace(buffer, STACKTRACE_SIZE);
+    std::cerr << "HANDLER CAUGHT signal Error: signal " << sig << std::endl;
+    symbols = backtrace_symbols(buffer, nptrs);
+    if (symbols == NULL) {
+        std::cerr << "Error getting backtrace symbols" << std::endl;
+    } else {
+        std::cerr << "C-LAND STACKTRACE: " << std::endl;
+        for (int i = 0; i < nptrs; i++) {
+            std::cerr << symbols[i] << std::endl;
+        }
+
+        _free(symbols);
+    }
+
+    // dump core if so configured
+    __sync_synchronize();
+    if (core_dump) {
+        core_dump(sig);
+    } else {
+        kill(getpid(), SIGABRT);
+    }
+}
+
 /*
  * C-land crash handling
  */
-
-DATADOG_AGENT_RTLOADER_API int handle_crashes(const rtloader_t *rtloader, const int enable)
+DATADOG_AGENT_RTLOADER_API int handle_crashes(const int enable, char **error)
 {
-    // enable implicit cast to bool
-    return AS_CTYPE(RtLoader, rtloader)->handleCrashes(enable) ? 1 : 0;
+
+    struct sigaction sa;
+    sa.sa_flags = SA_SIGINFO;
+    sa.sa_sigaction = signalHandler;
+
+    // on segfault - what else?
+    int err = sigaction(SIGSEGV, &sa, NULL);
+
+    if (enable && err == 0) {
+        __sync_synchronize();
+        core_dump = core;
+    }
+    if (err) {
+        std::ostringstream err_msg;
+        err_msg << "unable to set crash handler: " << strerror(errno);
+        *error = strdupe(err_msg.str().c_str());
+    }
+
+    return err;
 }
 #endif
